initial conversion to typescript/split files
