+++ .eslintrc.js
@@ -1,14 +1,15 @@
 module.exports = {
+  root: true,
   env: {
-    browser: true,
-    es6: true
+    node: true,
+    browser: true
   },
-  extends: ['eslint:recommended'],
-  globals: {},
-  parserOptions: {
-    ecmaVersion: 2018,
-    sourceType: 'module'
-  },
-  plugins: [],
-  rules: {}
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint'],
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+  rules: {
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
+    '@typescript-eslint/no-explicit-any': ['error', { ignoreRestArgs: true }],
+    '@typescript-eslint/no-non-null-assertion': 'off'
+  }
 }
