+++ meta.js
@@ -1,0 +1,10 @@
+// ==UserScript==
+// @name        Reddit Comments on Youtube
+// @description show reddit comments on youtube (and crunchyroll) videos
+// @namespace   RCOY
+// @version     0.2.9
+// @match       https://*.youtube.com/*
+// @match       https://*.crunchyroll.com/*
+// @match       https://www14.9anime.to/watch/*
+// @grant       none
+// ==/UserScript==
