+++ src/cmp/Comments.tsx
@@ -1,0 +1,7 @@
+interface Props {
+  comments: Comment[]
+}
+
+export const Comments = (props: Props) => {
+  return null
+}
