+++ src/constants.ts
@@ -1,0 +1,5 @@
+export const SCRIPT_NAME = 'RCOY'
+
+export const API_URL = 'https://www.reddit.com'
+
+export const COMMENT_LOAD_NUM = 20
